+++ 2015/src/common/utils/utils.h
@@ -7,5 +7,5 @@ char* file_to_string(const char*);
 void free_array(void**);
 char** file_to_array(const char*);
 
-int64_t min(uint8_t, ...);
-int64_t max(uint8_t, ...);
+int64_t min(int, ...);
+int64_t max(int, ...);
